% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomflowfield.R
\name{randomflowfield}
\alias{randomflowfield}
\title{Generate one direction of a random flow field}
\usage{
randomflowfield(n, m, std_dev)
}
\arguments{
\item{n}{Zonal (x) extent of the flow field.}

\item{m}{Meridional (y) extent of the flow field.}

\item{std_dev}{Standard deviation used for the random walk, see details.}
}
\value{
A matrix meant to be used as one direction of a random flow field.
}
\description{
Generate a matrix meant to be used as one direction of a random flow field.
}
\details{
This function proceeds in the following way: A starting value, either 1 or
-1, is drawn at random. This is assigned to the element in the top-left
corner of the matrix. The other values are then calculated as the sum of a
normal random variable with standard deviaton \code{std\_dev} and their left
neighbour (top row) or the mean of their left and their upper neighbour
(rest).
}
\examples{

U = randomflowfield(30,30,0.3)
V = randomflowfield(30,30,0.3)
quiver(U,V)
}
\seealso{
\code{\link{prepare.arena}}
}
\author{
Thorsten Lenser
}
\keyword{misc}
