\name{Sdata}
\docType{data}
\alias{Sdata}
\title{Polygon structure and land positions}
\description{
Example data set of a polygon structure over the Kerguelen plateau. 0 denotes land and positive integer values 
give the number of the polygon the respective grid point belongs to.
}
\usage{data(Sdata)}
\format{A matrix of integer values.}
\source{Thorsten Lenser, just an example.}
\keyword{datasets}