\name{Udata}
\docType{data}
\alias{Udata}
\title{EW component of a flow field}
\description{
Example data set of u-components of the flow field over the Kerguelen plateau. Derived from satellite measurements of sea surface height.
}
\usage{data(Udata)}
\format{Three columns, containing longitude, latitude and current speed in m/s. Unknown values are given as -9999.}
\source{Belinda Ronai, Australian Antarctic Division, unpublished data.}
\keyword{datasets}