\name{Vdata}
\docType{data}
\alias{Vdata}
\title{NS component of a flow field}
\description{
Example data set of v-components of the flow field over the Kerguelen plateau. Derived from satellite measurements of sea surface height.
}
\usage{data(Vdata)}
\format{Three columns, containing longitude, latitude and current speed in m/s. Unknown values are given as -9999.}
\source{Belinda Ronai, Australian Antarctic Division, unpublished data.}
\keyword{datasets}