% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomassTracking-package.R
\docType{data}
\name{Sdata}
\alias{Sdata}
\title{Polygon structure and land positions}
\format{A matrix of integer values.}
\source{
Thorsten Lenser, just an example.
}
\description{
Example data set of a polygon structure over the Kerguelen plateau. 0
denotes land and positive integer values give the number of the polygon the
respective grid point belongs to.
}
\keyword{datasets}
