% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomassTracking-package.R
\docType{data}
\name{Udata}
\alias{Udata}
\title{EW component of a flow field}
\format{Three columns, containing longitude, latitude and current speed in
m/s. Unknown values are given as -9999.}
\source{
Belinda Ronai, Australian Antarctic Division, unpublished data.
}
\description{
Example data set of u-components of the flow field over the Kerguelen
plateau. Derived from satellite measurements of sea surface height.
}
\keyword{datasets}
